+++ src/unpoly/history.js
@@ -6,7 +6,7 @@ The `up.history` module helps you work with the browser history.
 
 @see up.history.location
 @see up:location:changed
-@see a[up-back]
+@see [up-back]
 
 @see updating-history
 @see restoring-history
@@ -476,8 +476,8 @@ up.history = (function() {
   If no previous URL is known, the link will not be changed.
 
   > [NOTE]
-  > Clicking an `a[up-back]` will *not* call [`history.back()`](https://developer.mozilla.org/en-US/docs/Web/API/History/back).
-  > Instead the link's `[up-href]` attribute will be set to the actual, previous URL.
+  > Clicking an `[up-back]` link will *not* call [`history.back()`](https://developer.mozilla.org/en-US/docs/Web/API/History/back).
+  > Instead the link will [navigate](/up.navigate) to the previous URL.
 
   ### Example
 
@@ -492,12 +492,12 @@ up.history = (function() {
   ... will be transformed to:
 
   ```html
-  <a href="/default" up-href="/previous-page" up-scroll="restore" up-follow>
+  <a href="/default" up-follow up-href="/previous-page" up-scroll="restore" up-follow>
     Go back
   </a>
   ```
 
-  @selector a[up-back]
+  @selector [up-back]
   @stable
   */
   up.macro('[up-back]', function(link) {